+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This change log follows the format documented in [Keep a CHANGELOG].
 [semantic versioning]: http://semver.org/
 [keep a changelog]: http://keepachangelog.com/
 
-## v0.4.0 - 2026-08-??
+## v0.4.0 - 2026-08-06
 
 ### Added